fix mainPost
