+++ src/components/MainField/MainCharacter.tsx
@@ -1,74 +1,37 @@
 'use client'
-
+import CharacterSelect from '@/components/select/CharacterSelect'
 import { useCharacterInfoList } from '@/store/characterStore'
 import { useState, useEffect } from 'react'
 import Image from 'next/image'
 import Link from 'next/link'
 import Loading from '@image/icon/loading.svg'
 import { useSession } from 'next-auth/react'
-import Under from '@image/icon/under.svg'
 import UtileCharacterDataFetch from '@/components/utils/utilCharacterGet'
-
-interface CharacterInfo {
-  character_name: string
-  user_id: string
-  character_level: string
-  character_class: string
-  server_name: string
-  class_image: string
-  class_icon_url: string
-  transcendence: number
-  elixir: number
-  leap: number
-  enlightenment: number
-  evolution: number
-  disable: boolean
-}
 
 export default function MainCharacter() {
   const { data: session } = useSession()
-  const [mainNickName, setMainNickName] = useState('마이페이지-내정보-닉네임 설정')
+  const { characterInfo, setCharacterAllList } = useCharacterInfoList()
   const [loading, setLoading] = useState(false)
-  const [userId, setUserId] = useState('')
-  const [getCharacterList, setGetCharaacterList] = useState<CharacterInfo[]>([])
-
-  const { setCharacterInfo, characterInfo, setCharacterAllList, characterAllList } =
-    useCharacterInfoList()
-  const [hidden, setHidden] = useState(true)
-  const handler = (name: string) => {
-    const selectedCharacter = characterAllList.find((char) => char.character_name === name)
-    if (selectedCharacter) {
-      setCharacterInfo([selectedCharacter])
-      setHidden(!hidden)
-    }
-  }
-  const handlerHidden = () => {
-    setHidden(!hidden)
-  }
+  const [mainNickName, setMainNickName] = useState('마이페이지-내정보-닉네임 설정')
 
   useEffect(() => {
-    const characterHandelr = async () => {
-      const getCharacterList: CharacterInfo[] = await UtileCharacterDataFetch(userId)
-      setGetCharaacterList(getCharacterList)
-    }
     setLoading(true)
-    if (session && session.user.id) {
-      if (session.user.nickName != '') {
-        setMainNickName(session.user.nickName)
-        setUserId(session.user.id)
+    const fetchCharacterData = async () => {
+      if (session && session.user.id) {
+        session.user.nickName
+        const userId = session.user.id
+        const getCharacterList = await UtileCharacterDataFetch(userId) // await 추가
+        setCharacterAllList(getCharacterList)
+        if (session.user.nickName != '') {
+          setMainNickName(session.user.nickName)
+        }
+      } else {
       }
-      characterHandelr()
+      setLoading(false)
     }
-    setLoading(false)
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [session, userId])
 
-  useEffect(() => {
-    if (getCharacterList.length > 0) {
-      setCharacterInfo([characterAllList[0]])
-      setCharacterAllList(getCharacterList)
-    }
-  }, [characterAllList, getCharacterList, setCharacterAllList, setCharacterInfo])
+    fetchCharacterData() // 비동기 함수 호출
+  }, [session, setCharacterAllList])
 
   return (
     <>
@@ -77,7 +40,7 @@ export default function MainCharacter() {
           <Loading className='h-12 w-12 animate-spin text-white' />
         </div>
       )}
-      {userId != '' ? (
+      {session && session.user.id ? (
         <div className='flex h-full w-full flex-col'>
           <div className='flex items-center gap-2 text-white'>
             <Image
@@ -89,56 +52,7 @@ export default function MainCharacter() {
             />
             <span>{mainNickName}</span>
           </div>
-          <div className='relative mt-3 w-full'>
-            {characterInfo[0] ? (
-              <>
-                <button
-                  className='flex h-16 w-full items-center justify-between rounded-md border border-gray-300 px-1 shadow-md'
-                  onClick={handlerHidden}
-                >
-                  <div className='flex w-full items-center gap-4 overflow-hidden truncate whitespace-nowrap'>
-                    <div className='h-12 w-12 overflow-hidden rounded-full'>
-                      <Image
-                        src={characterInfo[0].class_image}
-                        alt='클래스 캐릭터 이미지'
-                        width={70}
-                        height={70}
-                        className='h-full w-full object-cover'
-                      />
-                    </div>
-                    <span className='overflow-hidden truncate whitespace-nowrap text-lg text-white'>
-                      {characterInfo[0].character_name}
-                    </span>
-                  </div>
-                  <Under className='h-4 w-4 text-white' />
-                </button>
-                <div
-                  className={`absolute left-0 top-full z-10 mt-1 w-full rounded-md bg-white shadow-md ${hidden ? 'hidden' : ''}`}
-                >
-                  {characterAllList.map((char) => (
-                    <div
-                      key={char.character_name}
-                      className='flex cursor-pointer items-center gap-4 p-2 hover:rounded-md hover:bg-gray-200'
-                      onClick={() => handler(char.character_name)}
-                    >
-                      <div className='h-12 w-12 overflow-hidden rounded-full'>
-                        <Image
-                          src={char.class_image}
-                          alt='클래스 캐릭터 이미지'
-                          width={70}
-                          height={70}
-                          className='h-full w-full object-cover'
-                        />
-                      </div>
-                      <span className='text-black'>{char.character_name}</span>
-                    </div>
-                  ))}
-                </div>
-              </>
-            ) : (
-              <span className='text-white'>캐릭터가 없습니다.</span>
-            )}
-          </div>
+          <CharacterSelect />
           {characterInfo[0] ? (
             <>
               <div className='mt-2 flex w-full flex-col items-center text-white md:flex-row md:gap-4 xl:mt-4 xl:w-full xl:flex-col'>
